+++ server/internal/store/sqlcdb/clusters.sql.go
@@ -27,7 +27,7 @@ SELECT $1::text, p.id, h.id, $2::text,
        $10::integer, $11::text,
        $12::integer,
        $13::boolean, $14::text,
-       $15::integer, $16::integer,
+       $15::bigint, $16::bigint,
        $17::bigint, $18::bigint,
        $19::bigint, $20::jsonb
 FROM projects p
@@ -60,8 +60,8 @@ type CreateClusterParams struct {
 	PgbouncerPublishPort          int32           `json:"pgbouncer_publish_port"`
 	PgbackrestEnabled             bool            `json:"pgbackrest_enabled"`
 	PgbackrestRepoPath            string          `json:"pgbackrest_repo_path"`
-	PgbackrestRetentionFull       int32           `json:"pgbackrest_retention_full"`
-	PgbackrestRetentionDiff       int32           `json:"pgbackrest_retention_diff"`
+	PgbackrestRetentionFull       int64           `json:"pgbackrest_retention_full"`
+	PgbackrestRetentionDiff       int64           `json:"pgbackrest_retention_diff"`
 	PgbackrestFullIntervalSeconds int64           `json:"pgbackrest_full_interval_seconds"`
 	PgbackrestDiffIntervalSeconds int64           `json:"pgbackrest_diff_interval_seconds"`
 	PgbackrestIncrIntervalSeconds int64           `json:"pgbackrest_incr_interval_seconds"`
@@ -493,8 +493,8 @@ type UpdateClusterParams struct {
 	PgbouncerPublishPort          int32           `json:"pgbouncer_publish_port"`
 	PgbackrestEnabled             bool            `json:"pgbackrest_enabled"`
 	PgbackrestRepoPath            string          `json:"pgbackrest_repo_path"`
-	PgbackrestRetentionFull       int32           `json:"pgbackrest_retention_full"`
-	PgbackrestRetentionDiff       int32           `json:"pgbackrest_retention_diff"`
+	PgbackrestRetentionFull       int64           `json:"pgbackrest_retention_full"`
+	PgbackrestRetentionDiff       int64           `json:"pgbackrest_retention_diff"`
 	PgbackrestFullIntervalSeconds int64           `json:"pgbackrest_full_interval_seconds"`
 	PgbackrestDiffIntervalSeconds int64           `json:"pgbackrest_diff_interval_seconds"`
 	PgbackrestIncrIntervalSeconds int64           `json:"pgbackrest_incr_interval_seconds"`
